FeaturePanel: new drawer
