+++ src/components/FeaturePanel/FeaturePanelInDrawer.tsx
@@ -1,9 +1,22 @@
 import React, { useState } from 'react';
 
-import { SwipeableDrawer, useTheme } from '@mui/material';
+import { SwipeableDrawer } from '@mui/material';
 import styled from 'styled-components';
 import { grey } from '@mui/material/colors';
-import { FeaturePanel } from './FeaturePanel';
+import { FeaturePanelInner } from './FeaturePanelInner';
+
+const Container = styled.div`
+  position: relative;
+  background: ${({ theme }) => theme.palette.background.paper};
+  margin-top: -86px;
+  border-top-left-radius: 12px;
+  border-top-right-radius: 12px;
+  visibility: visible;
+  right: 0;
+  left: 0;
+  box-shadow: 0 0 20px rgba(0, 0, 0, 0.2);
+  height: 100%;
+`;
 
 const Puller = styled.div`
   width: 30px;
@@ -12,18 +25,19 @@ const Puller = styled.div`
     theme.palette.mode === 'light' ? grey[300] : grey[900]};
   border-radius: 3px;
   position: absolute;
+  pointer-events: none;
   top: 8px;
   left: calc(50% - 15px);
 `;
 
 const ListContainer = styled('div')(() => ({
-  maxHeight: '90vh',
+  maxHeight: 'calc(100% - 8px)',
+  height: '100%',
   overflow: 'auto',
 }));
 
 export const FeaturePanelInDrawer = () => {
   const [open, setOpen] = useState(false);
-  const theme = useTheme();
 
   return (
     <SwipeableDrawer
@@ -33,29 +47,18 @@ export const FeaturePanelInDrawer = () => {
       onOpen={() => {
         setOpen(true);
       }}
-      swipeAreaWidth={72}
+      swipeAreaWidth={86}
       disableSwipeToOpen={false}
       ModalProps={{
         keepMounted: true,
       }}
     >
-      <div
-        style={{
-          position: 'relative',
-          background: theme.palette.background.paper,
-          marginTop: `-72px`,
-          borderTopLeftRadius: 8,
-          borderTopRightRadius: 8,
-          visibility: 'visible',
-          right: 0,
-          left: 0,
-        }}
-      >
+      <Container>
         <Puller />
         <ListContainer>
-          <FeaturePanel />
+          <FeaturePanelInner />
         </ListContainer>
-      </div>
+      </Container>
     </SwipeableDrawer>
   );
 };
