+++ src/components/utils/PanelHelpers.tsx
@@ -2,38 +2,30 @@ import React from 'react';
 import styled from 'styled-components';
 import { Scrollbars } from 'react-custom-scrollbars';
 import { useTheme } from '@mui/material';
-// import { isDesktop } from '../helpers';
+import { isDesktop } from '../helpers';
 import { useScrollShadow } from '../FeaturePanel/Climbing/utils/useScrollShadow';
-import { SEARCH_BOX_HEIGHT } from '../SearchBox/consts';
 
 // custom scrollbar
 // better: https://github.com/rommguy/react-custom-scroll
 // maybe https://github.com/malte-wessel/react-custom-scrollbars (larger)
-export const PanelWrapper = styled.div<{ fromLeft?: boolean }>`
-  /* position: absolute; */
-  /* left: 0; */
-  /* top: 72px; // TopPanel */
-  /* bottom: 0; */
+export const PanelWrapper = styled.div`
+  position: absolute;
+  left: 0;
+  top: 72px; // TopPanel
+  bottom: 0;
   background: ${({ theme }) => theme.palette.background.paper};
   color: ${({ theme }) => theme.palette.text.primary};
   overflow: hidden;
-  /* z-index: 1100; */
+  z-index: 1100;
   box-shadow: 0 0 20px rgba(0, 0, 0, 0.2);
 
   display: flex;
   flex-direction: column;
 
-  ${({ fromLeft }) =>
-    fromLeft
-      ? `
+  width: 100%;
+  @media ${isDesktop} {
     width: 410px;
-    height: 100%;
-    margin-top: 72px; // TopPanel
-  `
-      : `
-    width: 100vw;
-    height: 100%;
-  `}
+  }
 
   & > div > div {
     // disable pulling panel around on mobile
@@ -75,10 +67,7 @@ export const PanelScrollbars = ({ children }) => {
 export const PanelContent = styled.div`
   display: flex;
   flex-direction: column;
-  /* height: calc(
-    100vh - ${SEARCH_BOX_HEIGHT}px - 238px
-  ); // 100% - TopPanel - FeatureImage */
-  padding: 20px 0 0 0;
+  height: 100%;
 `;
 
 export const PanelFooter = styled.div`
